compiling

+++ module/Pequena/window.py
@@ -105,6 +105,8 @@ def check_for_modules(js_path):
 
 
 def handle_build_copy():
+    if not os.path.exists(client_dir):
+        return  # Compiled to exe
     copy_tree(client_dir, build_dir)
     script_str = "<script>window.addEventListener('pywebviewready', function () {\nconst PEQUENA = pywebview.api\n"
     new_html = ""
